backend/api/views.py
from rest_framework.decorators import api_view
from rest_framework.response import Response
from rest_framework import status
import joblib
import numpy as np
from .models import Race, TelemetryData
from ml_models.monte_carlo import MonteCarloRaceSimulator
import pandas as pd

# Load models
xgboost_model = joblib.load('ml_models/xgboost_winner_model.pkl')

@api_view(['GET'])
def live_data(request):
    """Get current live race data"""
    # Fetch latest telemetry
    latest_data = TelemetryData.objects.order_by('-timestamp')[:20]
    
    return Response({
        'positions': [...],
        'telemetry': [...],
        'predictions': {...}
    })

@api_view(['POST'])
def predict_race(request):
    """Run race prediction with XGBoost + Monte Carlo"""
    race_config = request.data
    
    # Prepare features
    features = np.array([[
        race_config['grid_position'],
        race_config['qualifying_time'],
        race_config['recent_avg_position'],
        # ... more features
    ]])
    
    # XGBoost prediction
    win_prob = xgboost_model.predict_proba(features)[0][1]
    
    # Monte Carlo simulation
    drivers_df = pd.DataFrame(race_config['drivers'])
    simulator = MonteCarloRaceSimulator(drivers_df)
    mc_results = simulator.run_simulation(n_simulations=5000)
    
    return Response({
        'xgboost_win_probability': float(win_prob),
        'monte_carlo_results': mc_results
    })

@api_view(['GET'])
def driver_stats(request, driver_id):
    """Get driver statistics and Alpha Score"""
    # Calculate Driver Alpha Score
    # (consistency + aggression + tire_efficiency)
    
    alpha_score = 0.75  # Placeholder
    
    return Response({
        'driver_id': driver_id,
        'alpha_score': alpha_score,
        'consistency': 0.80,
        'aggression': 0.70,
        'tire_efficiency': 0.85
    })

@api_view(['POST'])
def optimize_pit(request):
    """Get RL-optimized pit stop recommendation"""
    # Load RL agent and run inference
    
    recommended_lap = 28
    
    return Response({
        'recommended_pit_lap': recommended_lap,
        'confidence': 0.87
    })
