requirements.txt
Django==5.1
djangorestframework==3.15
django-cors-headers==4.3
celery[redis]==5.4
channels==4.0
channels-redis==4.2
psycopg2-binary==2.9
fastf1==3.3
xgboost==2.0
tensorflow==2.15  # or torch
pmdarima==2.0
scikit-learn==1.4
pandas==2.2
numpy==1.26
gunicorn==21.2
